root: Extract geography and JSON modules
Extract geography and JSON classes into their own respective modules.
Also add a couple of tests to ensure these modules are compliant with
the project's coverage threshold.
