+++ settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = "Journal3"
 include ':app-android-journal3'
 include ':app-kmm-journal3'
+include ':lib-kmm-geography'
+include ':lib-kmm-json'
 include ':lib-kmm-foundation'
